. modules are now aware of the sampling rate . doubles can be now used in the mdl's data structures . fixes to ondemand querying . adding code to profile capture, disabled by default

+++ src/branches/2.0/modules/autofocus/export.cs
@@ -182,7 +182,7 @@ namespace CoMo.Modules.autofocus
             }
 
             foreach(AutofocusTuple t in tuples)
-                update_tree(t.addr, t.bytes);
+                update_tree(t.addr, (ulong) t.bytes);
         }
     }
 }
